[BE] refactor(#114): 테스트 코드 리펙토링
