+++ backend/src/test/java/com/example/backend/auth/TestConfig.java
@@ -42,85 +42,5 @@ public class TestConfig {
     public static String createAuthorizationHeader(String accessToken, String refreshToken) {
         return BEARER + " " + accessToken + " " + refreshToken;
     }
-
-    // StudyInfo 생성 해주는 메소드
-    public static StudyInfo generateStudyInfo(Long userId) {
-        return StudyInfo.builder()
-                .userId(userId)
-                .topic("Sample Study")
-                .score(100)
-                .endDate(LocalDate.now().plusMonths(3))
-                .info("This is a sample study.")
-                .status(StudyStatus.STUDY_PUBLIC)
-                .joinCode("ABC123")
-                .maximumMember(5)
-                .currentMember(3)
-                .lastCommitDay(LocalDate.now())
-                .profileImageUrl("https://example.com/profile.jpg")
-                .notice("Important notice for the study.")
-                .repositoryInfo(new RepositoryInfo("구영민", "aaa333", "BRANCH_NAME"))
-                .periodType(StudyPeriodType.STUDY_PERIOD_EVERYDAY)
-                .build();
-    }
-
-    // generateStudyInfoRegisterRequest 생성 해주는 메소드
-    public static StudyInfoRegisterRequest generateStudyInfoRegisterRequest(Long userId) {
-        List<StudyCategory> categories = new ArrayList<>();
-        categories.add(new StudyCategory("c++"));
-        categories.add(new StudyCategory("python"));
-
-        return StudyInfoRegisterRequest.builder()
-                .userId(userId)
-                .topic("Sample Study")
-                .endDate(LocalDate.now().plusMonths(3))
-                .info("This is a sample study.")
-                .status(StudyStatus.STUDY_PUBLIC)
-                .maximumMember(5)
-                .profileImageUrl("https://example.com/profile.jpg")
-                .repositoryInfo(new RepositoryInfo("구영민", "aaa333", "BRANCH_NAME"))
-                .periodType(StudyPeriodType.STUDY_PERIOD_EVERYDAY)
-                .categories(categories)
-                .build();
-    }
-
-    // MaximumMember가 10보다 클 때, generateStudyInfoRegisterRequest 생성 해주는 메소드
-    public static StudyInfoRegisterRequest generateStudyInfoRegisterRequestWhenMaximumMemberExceed10(Long userId) {
-        List<StudyCategory> categories = new ArrayList<>();
-        categories.add(new StudyCategory("c++"));
-        categories.add(new StudyCategory("python"));
-
-        return StudyInfoRegisterRequest.builder()
-                .userId(userId)
-                .topic("Sample Study")
-                .endDate(LocalDate.now().plusMonths(3))
-                .info("This is a sample study.")
-                .status(StudyStatus.STUDY_PUBLIC)
-                .maximumMember(11)
-                .profileImageUrl("https://example.com/profile.jpg")
-                .repositoryInfo(new RepositoryInfo("구영민", "aaa333", "BRANCH_NAME"))
-                .periodType(StudyPeriodType.STUDY_PERIOD_EVERYDAY)
-                .categories(categories)
-                .build();
-    }
-
-    // MaximumMember가 1보다 작을 때, generateStudyInfoRegisterRequest 생성 해주는 메소드
-    public static StudyInfoRegisterRequest generateStudyInfoRegisterRequestWhenMaximumMemberLessThan1(Long userId) {
-        List<StudyCategory> categories = new ArrayList<>();
-        categories.add(new StudyCategory("c++"));
-        categories.add(new StudyCategory("python"));
-
-        return StudyInfoRegisterRequest.builder()
-                .userId(userId)
-                .topic("Sample Study")
-                .endDate(LocalDate.now().plusMonths(3))
-                .info("This is a sample study.")
-                .status(StudyStatus.STUDY_PUBLIC)
-                .maximumMember(-1)
-                .profileImageUrl("https://example.com/profile.jpg")
-                .repositoryInfo(new RepositoryInfo("구영민", "aaa333", "BRANCH_NAME"))
-                .periodType(StudyPeriodType.STUDY_PERIOD_EVERYDAY)
-                .categories(categories)
-                .build();
-    }
 }
